Remove cgroup v1 plugins
