+++ components/ws-daemon/pkg/daemon/daemon.go
@@ -76,11 +76,6 @@ func NewDaemon(config Config, reg prometheus.Registerer) (*Daemon, error) {
 		return nil, err
 	}
 
-	cgroupV1IOLimiter, err := cgroup.NewIOLimiterV1(config.IOLimit.WriteBWPerSecond.Value(), config.IOLimit.ReadBWPerSecond.Value(), config.IOLimit.WriteIOPS, config.IOLimit.ReadIOPS)
-	if err != nil {
-		return nil, err
-	}
-
 	cgroupV2IOLimiter, err := cgroup.NewIOLimiterV2(config.IOLimit.WriteBWPerSecond.Value(), config.IOLimit.ReadBWPerSecond.Value(), config.IOLimit.WriteIOPS, config.IOLimit.ReadIOPS)
 	if err != nil {
 		return nil, err
@@ -93,9 +88,7 @@ func NewDaemon(config Config, reg prometheus.Registerer) (*Daemon, error) {
 
 	cgroupPlugins, err := cgroup.NewPluginHost(config.CPULimit.CGroupBasePath,
 		&cgroup.CacheReclaim{},
-		&cgroup.FuseDeviceEnablerV1{},
 		&cgroup.FuseDeviceEnablerV2{},
-		cgroupV1IOLimiter,
 		cgroupV2IOLimiter,
 		&cgroup.ProcessPriorityV2{
 			ProcessPriorities: map[cgroup.ProcessType]int{
@@ -143,7 +136,6 @@ func NewDaemon(config Config, reg prometheus.Registerer) (*Daemon, error) {
 
 	var configReloader CompositeConfigReloader
 	configReloader = append(configReloader, ConfigReloaderFunc(func(ctx context.Context, config *Config) error {
-		cgroupV1IOLimiter.Update(config.IOLimit.WriteBWPerSecond.Value(), config.IOLimit.ReadBWPerSecond.Value(), config.IOLimit.WriteIOPS, config.IOLimit.ReadIOPS)
 		cgroupV2IOLimiter.Update(config.IOLimit.WriteBWPerSecond.Value(), config.IOLimit.ReadBWPerSecond.Value(), config.IOLimit.WriteIOPS, config.IOLimit.ReadIOPS)
 		procV2Plugin.Update(config.ProcLimit)
 		if config.NetLimit.Enabled {
